Fix type casting errors
 + cast to __m128i* when calling _mm_loadu_si128/_mm_storeu_si128
 + change __m256i to __m128i when calling _mm_storeu_si128. change __m128i to __m256i when calling _mm256_loadu_si256
 + change __m128i to __m256i when calling _mm256_storeu_si256
 + cast to `_Float16` from `float`

+++ lib/ext/bg/bg.c
@@ -129,7 +129,7 @@ unsigned char* BG_compress_args(int dataType, void *data, size_t *outSize, int b
 				return NULL;
 			}
 
-			u64_ptr=(unsigned long int*)data_;
+			u64_ptr=(unsigned long long int*)data_;
 			/* Create mask */
 			msk_f64_u64_zro=0ul; /* Zero all bits */
 			msk_f64_u64_zro=~msk_f64_u64_zro; /* Turn all bits to ones */
